Heroku is not binding to the port but the port wasnt
hard-coded. I changed the logic to grab the port from the env
at the time the function in called rather than from the
const declaration at the top of the file. Perhaps it was undefined
when the constant was created and therefore we tried to assign a
number. The heroku logs dont seem to support this, idea I saw
an arbitrary port number and not 3000 or 3005. Everything works
just fine locally...